+++ sort.go
@@ -2,11 +2,10 @@ package jsonvalue
 
 import (
 	"bytes"
+	"io"
 	"sort"
 	"strconv"
 	"strings"
-
-	"github.com/Andrew-M-C/go.jsonvalue/internal/buffer"
 )
 
 // ---------------- array sorting ----------------
@@ -186,7 +185,7 @@ func DefaultStringSequence(parent *ParentInfo, key1, key2 string, v1, v2 *V) boo
 	return strings.Compare(key1, key2) <= 0
 }
 
-func (sov *sortObjectV) marshalObjectWithLessFunc(buf buffer.Buffer, opt *Opt) {
+func (sov *sortObjectV) marshalObjectWithLessFunc(buf io.Writer, opt *Opt) {
 	// sort
 	sort.Sort(sov)
 
@@ -235,7 +234,7 @@ func newSortObjectV(v *V, parentInfo *ParentInfo, opt *Opt) *sortObjectV {
 }
 
 // marshalObjectWithStringSlice use a slice to determine sequence of object
-func (sssv *sortStringSliceV) marshalObjectWithStringSlice(buf buffer.Buffer, opt *Opt) {
+func (sssv *sortStringSliceV) marshalObjectWithStringSlice(buf io.Writer, opt *Opt) {
 	// sort
 	sort.Sort(sssv)